Activated more output

+++ scripts/sync_ldap_to_irods.py
@@ -34,7 +34,7 @@ class SyncRunner(object):
         # Create users
         for user in self.users:
             if not irods.user_exists(user.username):
-                #sys.stderr.write("User %s missing, so creating now ...\n" % user.username)
+                sys.stderr.write("User %s missing, so creating now ...\n" % user.username)
                 irods.create_user(user.username, usertype="rodsuser")
                 
         # Get all groups
@@ -43,13 +43,13 @@ class SyncRunner(object):
         # Create groups
         for group in groups:
             if not irods.group_exists(group.groupname):
-                #sys.stderr.write("Group %s missing, so creating now ...\n" % group.groupname)
+                sys.stderr.write("Group %s missing, so creating now ...\n" % group.groupname)
                 irods.create_group(group.groupname)
 
         # Connect users and groups
         for group in groups:
             for username in group.usernames:
-                if irods.user_exists(username):
+                if irods.user_exists(username) and not irods.group_has_user(group.groupname, username):
                     # sys.stderr.write("Now adding user %s to group %s ...\n" % (username,group.groupname))
                     irods.add_user_to_group(username, group.groupname)
                     
@@ -194,16 +194,23 @@ class IRodsConnector(object):
         exec_cmd(cmd)
         
     def delete_group(self, groupname):
-        if self.group_exists(groupname):
-            cmd = self.get_iadmin_p() + " rmgroup " + groupname
-            exec_cmd(cmd)
+        cmd = self.get_iadmin_p() + " rmgroup " + groupname
+        exec_cmd(cmd)
+        
+    def group_has_user(self, group, user):
+        cmd = self.get_iadmin_p() + " lg " + group
+        output = exec_cmd(cmd)
+        if user + "#" in output:
+            return True
         else:
-            #sys.stderr.write("Group does not existing, so can not delete: %s\n" % groupname)
-            pass
+            return False
             
     def add_user_to_group(self, username, groupname):
-        cmd = self.get_iadmin_p() + " atg %s %s" % (groupname, username)
-        exec_cmd(cmd)
+        if not username and not groupname:
+            cmd = self.get_iadmin_p() + " atg %s %s" % (groupname, username)
+            exec_cmd(cmd)
+        else:
+            print("Username %s or group %s is empty!" % (username, groupname))
         
     def folder_exists(self, folder):
         cmd = self.get_ils_p() + " " + folder
@@ -294,15 +301,15 @@ class TestSyncRunner(object):
         irods = IRodsConnector()
         for user in irods.list_users_in_zone("ssUppnexZone"):
             if not "rods" in user:
-                #sys.stderr.write("Now deleting user " + user + "...\n")
+                sys.stderr.write("Now deleting user " + user + "...\n")
                 irods.delete_user(user)
 
     @classmethod    
     def delete_all_groups(self):
         irods = IRodsConnector()
         for group in self.syncrunner.get_groups():
-            if not "public" in group.groupname and not "rodsadmin" in group.groupname:
-                #sys.stderr.write("Now deleting group " + group.groupname + "...\n")
+            if not "public" in group.groupname and not "rodsadmin" in group.groupname and irods.group_exists(group.groupname):
+                sys.stderr.write("Now deleting group " + group.groupname + "...\n")
                 try:
                     irods.delete_group(group.groupname)
                 except:
